fix: extensive loading harmonization config
When `Event` is init, it automatically loads the harmonization config
file. This means that for every event uploaded, the entire harmonization
config file is loaded and parsed. This dramatically extends the time
needed to parse a single event.

This PR ensures that the harmonization file is only loaded once and then
used for generating all `Event` objects.

Fixes: #79